requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gradio>=5.6.0
httpx>=0.27.0
pyyaml>=6.0
pydantic>=2.9.0
numpy>=1.26.0
Pillow>=10.4.0
opencv-python>=4.10.0
networkx>=3.4
scikit-learn>=1.5.0
pytest>=8.3.0
